REFACTOR: Re-modularize utils directory.
