+++ src/getters/getSimplePrice.ts
@@ -1,14 +1,14 @@
 require("dotenv").config({ path: '../../.env' });
 const rp = require('request-promise');
-const uploadS3 = require('./s3');
-const dateUtils = require('./dateUtils');
+const uploadS3 = require('../s3');
+const dateUtils = require('../dateUtils');
 
-async function getTokenPrice(symbol, balance, tokenAddress) {
+async function getSimplePrice(symbol:string, balance:number) {
   const requestOptions = {
     method: 'GET',
-    uri: 'https://api.coingecko.com/api/v3/simple/token_price/ethereum',
+    uri: 'https://api.coingecko.com/api/v3/simple/price',
     qs: {
-      'contract_addresses': tokenAddress,
+      'ids': 'ethereum',
       'vs_currencies': 'USD',
       'include_market_cap': 'true',
       'include_24hr_vol': 'true',
@@ -18,32 +18,33 @@ async function getTokenPrice(symbol, balance, tokenAddress) {
     json: true,
     gzip: true
   };
-  await rp(requestOptions).then(response => {
-    let tokenJson = response[`${tokenAddress}`.toLocaleLowerCase()]
+  await rp(requestOptions).then((response:any) => {
+    // console.log('symbol=%s, balance=%s, response=%s', symbol, balance, JSON.stringify(response))
+    let tokenJson = response.ethereum
     let price = tokenJson.usd.toFixed(2)
     let volume = tokenJson.usd_24h_vol.toFixed(2)
     console.log('symbol=%s, balance=%s, price=%s, usd_24h_vol=%s', symbol, balance, price, volume)
 
     const fileName = getFileName(symbol)
-    const content = getContent(symbol, tokenAddress, balance, price, volume);
+    const content = getContent(symbol, balance, price, volume);
     uploadS3.uploadFile(fileName, JSON.stringify(content));
-  }).catch((err) => {
+  }).catch((err:any) => {
     console.log('API call error:', err.message);
   });
 }
 
-function getFileName(token) {
+function getFileName(token:string) {
   const year = dateUtils.getYear()
   const month = dateUtils.getMonth()
   const day = dateUtils.getDay()
   return `${token}-${year}-${month}-${day}.json`;
 }
 
-function getContent(token, tokenAddress, quantity, price, volume) {
+function getContent(token:string, quantity:number, price:number, volume:number) {
   const year = dateUtils.getYear()
   const month = dateUtils.getMonth()
   const day = dateUtils.getDay()
-  const priceVolumeSource = `https://api.coingecko.com/api/v3/simple/token_price/ethereum?contract_addresses=${tokenAddress}&vs_currencies=USD&include_market_cap=true&include_24hr_vol=true&include_24hr_change=true&include_last_updated_at=true`
+  const priceVolumeSource = `https://api.coingecko.com/api/v3/simple/price?ids=ethereum&vs_currencies=USD&include_market_cap=true&include_24hr_vol=true&include_24hr_change=true&include_last_updated_at=true`
   return {
     yyyy: year,
     mm: month,
@@ -58,6 +59,6 @@ function getContent(token, tokenAddress, quantity, price, volume) {
   }
 }
 
-module.exports = {
-  getTokenPrice: getTokenPrice
+export {
+  getSimplePrice
 }
